Refactored unit tests to TypeScript
