+++ spec/support/jasmine.json
@@ -1,7 +1,7 @@
 {
-  "spec_dir": "spec",
+  "spec_dir": "",
   "spec_files": [
-    "**/*[sS]pec.js"
+    "dist/cjs/**/*[sS]pec.js"
   ],
   "helpers": [
     "helpers/**/*.js"
