+++ src/cuid.spec.ts
@@ -1,5 +1,5 @@
-const cuid = require('../dist/cjs/cuid').default;
-const hiresCuid = require('../dist/cjs/cuid/hires').default;
+import cuid from './cuid';
+import hiresCuid from './cuid/hires';
 
 describe('cuid', function() {
     it ('high resolution', () => {
@@ -11,7 +11,7 @@ describe('cuid', function() {
 
         // Generate hires cuids
         const n = 100000;
-        let cuids = new Array(n);
+        const cuids = new Array(n) as string[];
         for (let i = 0; i < n; i++) {
             cuids[i] = hiresCuid();
         }
@@ -19,7 +19,7 @@ describe('cuid', function() {
         // Expect all first 11 chars (7 ms + 4 ns) to be different and lexicographically sortable (cuid[n] < cuid[n+1])
         for (let i = 1; i < n; i++) {
             const t1 = cuids[i-1].slice(1, 12), t2 = cuids[i].slice(1, 12);
-            expect(t1).toBeLessThan(t2);
+            expect(t1 < t2).toBeTrue();
         }
 
         // debugger;
